requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
elasticsearch==7.9.1
Flask==1.1.2
fuzzywuzzy==0.18.0
gast
gensim==3.8.3
gevent==20.9.0
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.33.2
h5py==3.0.0
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
nltk==3.5
numpy==1.18.5
protobuf==3.13.0
python-Levenshtein==0.12.0
regex==2020.10.28
requests==2.24.0
scipy==1.4.1
six==1.15.0
smart-open==3.0.0
tensorboard
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.4
termcolor==1.1.0
textblob==0.15.3
tqdm==4.51.0
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2
